# docs/openapi/mymall.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "mymall 商城 API",
    "description": "宠物商城 REST API（user + catalog + order 合并文档）。\n\n统一响应 `{code, msg, data}`，code=200 表示业务成功。\n经 APISIX 网关: http://localhost:9080\n本地直连: user :8881, catalog :8882, order :8883",
    "version": "1.0.0",
    "contact": {
      "name": "mymall"
    }
  },
  "host": "localhost:9080",
  "basePath": "/",
  "paths": {
    "/api/v1/user/login": {
      "post": {
        "description": "手机号 + 密码登录，返回 JWT Token",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "用户登录",
        "parameters": [
          {
            "description": "登录参数",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.LoginReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "登录成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.LoginResp"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/user/profile": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "需要 JWT 鉴权",
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "获取个人资料",
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.UserProfileResp"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/user/register": {
      "post": {
        "description": "手机号注册新用户",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "用户注册",
        "parameters": [
          {
            "description": "注册参数",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.RegisterReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "注册成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.UserInfo"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/product_category/detail": {
      "get": {
        "description": "根据分类 ID 查询详情",
        "produces": [
          "application/json"
        ],
        "tags": [
          "分类"
        ],
        "summary": "分类详情",
        "parameters": [
          {
            "type": "integer",
            "description": "分类 ID",
            "name": "id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.CategoryDetailResp"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/product_category/list": {
      "get": {
        "description": "分页查询商品分类",
        "produces": [
          "application/json"
        ],
        "tags": [
          "分类"
        ],
        "summary": "分类列表",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 10,
            "description": "每页数量",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.CategoryListResp"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/products/detail": {
      "get": {
        "description": "根据商品 ID 查询详情",
        "produces": [
          "application/json"
        ],
        "tags": [
          "商品"
        ],
        "summary": "商品详情",
        "parameters": [
          {
            "type": "integer",
            "description": "商品 ID",
            "name": "id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.ProductDetail"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/products/list": {
      "get": {
        "description": "分页查询在售商品",
        "produces": [
          "application/json"
        ],
        "tags": [
          "商品"
        ],
        "summary": "商品列表",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 10,
            "description": "每页数量",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.ProductListResp"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/orders": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "分页查询当前用户订单",
        "produces": [
          "application/json"
        ],
        "tags": [
          "订单"
        ],
        "summary": "我的订单列表",
        "parameters": [
          {
            "type": "integer",
            "description": "用户 ID（网关注入）",
            "name": "X-User-Id",
            "in": "header"
          },
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 10,
            "description": "每页数量",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrderListResp"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "需要 JWT；经 APISIX 网关时从 Authorization 注入 X-User-Id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订单"
        ],
        "summary": "创建订单",
        "parameters": [
          {
            "type": "integer",
            "description": "用户 ID（网关注入，直连调试时必填）",
            "name": "X-User-Id",
            "in": "header"
          },
          {
            "description": "下单商品",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateOrderReq"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "创建成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrderInfo"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订单"
        ],
        "summary": "订单详情",
        "parameters": [
          {
            "type": "integer",
            "description": "用户 ID（网关注入）",
            "name": "X-User-Id",
            "in": "header"
          },
          {
            "type": "integer",
            "description": "订单 ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrderInfo"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}/cancel": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "pending/confirmed 状态可取消，会触发库存回滚",
        "produces": [
          "application/json"
        ],
        "tags": [
          "订单"
        ],
        "summary": "取消订单",
        "parameters": [
          {
            "type": "integer",
            "description": "用户 ID（网关注入）",
            "name": "X-User-Id",
            "in": "header"
          },
          {
            "type": "integer",
            "description": "订单 ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "取消成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/apidoc.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/apidoc.EmptyData"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    }
  },
  "definitions": {
    "apidoc.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 200
        },
        "data": {},
        "msg": {
          "type": "string",
          "example": "success"
        }
      }
    },
    "dto.LoginReq": {
      "type": "object",
      "required": [
        "mobile",
        "password"
      ],
      "properties": {
        "mobile": {
          "type": "string",
          "example": "13800138000"
        },
        "password": {
          "type": "string",
          "minLength": 6,
          "example": "123456"
        }
      }
    },
    "dto.LoginResp": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        },
        "user": {
          "$ref": "#/definitions/dto.UserInfo"
        }
      }
    },
    "dto.RegisterReq": {
      "type": "object",
      "required": [
        "mobile",
        "password"
      ],
      "properties": {
        "mobile": {
          "type": "string",
          "example": "13800138000"
        },
        "password": {
          "type": "string",
          "minLength": 6,
          "example": "123456"
        }
      }
    },
    "dto.UserInfo": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string",
          "example": ""
        },
        "created_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        },
        "gender": {
          "type": "integer",
          "example": 0
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "mobile": {
          "type": "string",
          "example": "13800138000"
        },
        "nickname": {
          "type": "string",
          "example": "用户8000"
        },
        "status": {
          "type": "integer",
          "example": 1
        },
        "updated_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        }
      }
    },
    "dto.UserProfileResp": {
      "type": "object",
      "properties": {
        "avatar": {
          "type": "string",
          "example": ""
        },
        "created_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        },
        "gender": {
          "type": "integer",
          "example": 0
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "mobile": {
          "type": "string",
          "example": "13800138000"
        },
        "nickname": {
          "type": "string",
          "example": "用户8000"
        },
        "status": {
          "type": "integer",
          "example": 1
        },
        "updated_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        }
      }
    },
    "dto.CategoryDetailResp": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        },
        "description": {
          "type": "string"
        },
        "icon": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "image": {
          "type": "string"
        },
        "is_show": {
          "type": "boolean",
          "example": true
        },
        "level": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "猫粮"
        },
        "parent_id": {
          "type": "integer",
          "example": 0
        },
        "product_count": {
          "type": "integer",
          "example": 10
        },
        "sort_order": {
          "type": "integer",
          "example": 0
        }
      }
    },
    "dto.CategoryItem": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        },
        "description": {
          "type": "string"
        },
        "icon": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "image": {
          "type": "string"
        },
        "is_show": {
          "type": "boolean",
          "example": true
        },
        "level": {
          "type": "integer",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "猫粮"
        },
        "parent_id": {
          "type": "integer",
          "example": 0
        },
        "product_count": {
          "type": "integer",
          "example": 10
        },
        "sort_order": {
          "type": "integer",
          "example": 0
        }
      }
    },
    "dto.CategoryListResp": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dto.CategoryItem"
          }
        },
        "page": {
          "type": "integer",
          "example": 1
        },
        "page_size": {
          "type": "integer",
          "example": 10
        },
        "total": {
          "type": "integer",
          "example": 20
        },
        "total_page": {
          "type": "integer",
          "example": 2
        }
      }
    },
    "dto.ProductDetail": {
      "type": "object",
      "properties": {
        "category_id": {
          "type": "integer",
          "example": 1
        },
        "created_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "main_image": {
          "type": "string"
        },
        "market_price": {
          "type": "number"
        },
        "name": {
          "type": "string",
          "example": "猫粮 10kg"
        },
        "pet_type": {
          "type": "string",
          "example": "cat"
        },
        "product_no": {
          "type": "string",
          "example": "P20260001"
        },
        "sale_price": {
          "type": "number",
          "example": 159
        },
        "status": {
          "type": "string",
          "example": "on_sale"
        },
        "stock": {
          "type": "integer",
          "example": 100
        },
        "subtitle": {
          "type": "string"
        }
      }
    },
    "dto.ProductListItem": {
      "type": "object",
      "properties": {
        "category_id": {
          "type": "integer",
          "example": 1
        },
        "created_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        },
        "discount": {
          "type": "number",
          "example": 100
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "is_hot": {
          "type": "boolean",
          "example": true
        },
        "is_new": {
          "type": "boolean",
          "example": false
        },
        "is_recommend": {
          "type": "boolean",
          "example": true
        },
        "main_image": {
          "type": "string",
          "example": "https://example.com/img.jpg"
        },
        "market_price": {
          "type": "number",
          "example": 199
        },
        "name": {
          "type": "string",
          "example": "猫粮 10kg"
        },
        "pet_type": {
          "type": "string",
          "example": "cat"
        },
        "product_no": {
          "type": "string",
          "example": "P20260001"
        },
        "sale_price": {
          "type": "number",
          "example": 159
        },
        "sold_count": {
          "type": "integer",
          "example": 10
        },
        "status": {
          "type": "string",
          "example": "on_sale"
        },
        "stock": {
          "type": "integer",
          "example": 100
        },
        "subtitle": {
          "type": "string",
          "example": "全年龄段"
        }
      }
    },
    "dto.ProductListResp": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dto.ProductListItem"
          }
        },
        "total": {
          "type": "integer",
          "example": 100
        }
      }
    },
    "apidoc.EmptyData": {
      "type": "object"
    },
    "dto.CreateOrderItem": {
      "type": "object",
      "required": [
        "product_id",
        "quantity"
      ],
      "properties": {
        "product_id": {
          "type": "integer",
          "example": 1
        },
        "quantity": {
          "type": "integer",
          "minimum": 1,
          "example": 2
        }
      }
    },
    "dto.CreateOrderReq": {
      "type": "object",
      "required": [
        "items"
      ],
      "properties": {
        "items": {
          "type": "array",
          "minItems": 1,
          "items": {
            "$ref": "#/definitions/dto.CreateOrderItem"
          }
        }
      }
    },
    "dto.OrderInfo": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dto.OrderItem"
          }
        },
        "order_no": {
          "type": "string",
          "example": "ORDabc12345"
        },
        "status": {
          "type": "string",
          "enum": [
            "pending",
            "confirmed",
            "failed",
            "cancelled"
          ],
          "example": "pending"
        },
        "total_amount": {
          "type": "number",
          "example": 318
        },
        "updated_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        },
        "user_id": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "dto.OrderItem": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string",
          "example": "2026-01-01 12:00:00"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "order_id": {
          "type": "integer",
          "example": 1
        },
        "price": {
          "type": "number",
          "example": 159
        },
        "product_id": {
          "type": "integer",
          "example": 1
        },
        "product_name": {
          "type": "string",
          "example": "猫粮 10kg"
        },
        "quantity": {
          "type": "integer",
          "example": 2
        }
      }
    },
    "dto.OrderListResp": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dto.OrderInfo"
          }
        },
        "total": {
          "type": "integer",
          "example": 10
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "JWT Token，格式: Bearer {token}",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "tags": [
    {
      "name": "用户",
      "description": "注册、登录、个人资料"
    },
    {
      "name": "商品",
      "description": "商品列表与详情"
    },
    {
      "name": "分类",
      "description": "商品分类"
    },
    {
      "name": "订单",
      "description": "下单、查单、取消"
    }
  ],
  "schemes": [
    "http"
  ]
}